D  .scribe/.filehashes D  .scribe/auth.md D  .scribe/endpoints.cache/00.yaml D  .scribe/endpoints/00.yaml D  .scribe/endpoints/custom.0.yaml D  .scribe/intro.md M  app/Http/Controllers/FavoriteProduct/FavoriteProductDestroyController.php M  app/Http/Controllers/FavoriteProduct/FavoriteProductIndexController.php M  app/Http/Controllers/Order/OrderIndexController.php M  app/Http/Controllers/Order/OrderShowController.php M  app/Http/Controllers/Product/ProductIndexController.php D  public/docs/collection.json D  public/docs/css/theme-default.print.css D  public/docs/css/theme-default.style.css D  public/docs/images/navbar.png D  public/docs/index.html D  public/docs/js/theme-default-4.35.0.js D  public/docs/js/tryitout-4.35.0.js D  public/docs/openapi.yaml M  routes/api.php

+++ app/Http/Controllers/FavoriteProduct/FavoriteProductDestroyController.php
@@ -20,8 +20,12 @@ class FavoriteProductDestroyController extends Controller
     public function __invoke(Request $request, $id)
     {
         $favoriteProduct = FavoriteProduct::findOrFail($id);
-        $favoriteProduct->delete();
 
-        return $this->response([], 'Любимый продукт успешно удален!');
+        if ($favoriteProduct->user_id == $request->user()->id){
+            $favoriteProduct->delete();
+            return $this->response([], 'Любимый продукт успешно удален!');
+        }
+
+        return $this->response([], 'Вы не имеете право удалять чужой любимый продукт!', 403);
     }
 }
